got data submitting working for everything but turn 1000

+++ DataMine/database/migrations/2016_06_24_022004_create_ziptopias_table.php
@@ -14,7 +14,7 @@ class CreateZiptopiasTable extends Migration
     {
         Schema::create('ziptopias', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('client_id')->unsigned();
+            $table->double('client_id', 13, 0)->unsigned();
             $table->integer('start_time')->unsigned();
             $table->integer('end_time')->unsigned()->nullable();
             $table->timestamps();
